WeaponsCardWeaponStats and WeaponsModalSelection moved to respective container folders
